+++ src/UIComponents/molecules/BlockerScreen/index.js
@@ -1,0 +1,2 @@
+import comp from "./BlockerScreen";
+export default comp;